This is a half-baked effort to get the unit tests working with nunit for mono compatibility.
The tests have been copied into a new project "Nunit" because the Test
project isn't recognized by MonoDevelop.  The test attributes have been
changed to be recognized by Nunit.

The tests compile and run, but there are many failures, indicating
library incompatibility. Also, the tests are not recognized by the
MonoDevelop UI - I only managed to get them to run from the console.

Once the library issue is figured out, I'll have to
de-duplicate the tests between the two test projects.

I'll check this in to a branch and try to continue the work for mono
compatibility in Windows.